CTP-3557 Migrate editor dependencies to Tiny
This plugin previously depended on editor_atto for the feedback form's
editor.  I'm not sure why an editor subclass is used (possibly due to
the way the feedback dialog was rendered) so have just migrated this to
the default editor for Moodle 4.4, editor_tiny, regardless of the
user's preferred editor.

There were also dependencies on editor_tinymce, which was removed in
Moodle 4.2.  These dependencies didn't appear to have any functional
impact and have been removed.