Fix: update_status() is deprecated
With debugging enabled, a plagiarism plugin installed, and
$CFG->enableplagiarism = 1 a teacher would see the following message at
the top of mod/coursework/view.php?id=<ID>:

  plagiarism_plugin::update_status() is deprecated. Use
  plagiarism_turnitin_before_standard_top_of_body_html() instead

The call to plagiarism_update_status() has been removed.  As far as I
can see this had no effect on this page when plagiarism_turnitin was
installed except to include a hidden "We’ve noticed high latency in your
connection ..." message.