CTP-4858 General feedback
Code tidy-up for the "General feedback" field for the revised user
interface.

Also, refactor Mustache templates and renderer function for intro area
so that <hr> dividers are only shown when there's content to separate.

Also, fix existing issue where if "General feedback release date" is
set the general feedback is never shown to the student.