CTP-5437 Hide profile pictures for blind marking
Hide user profile pictures when blind marking enabled (regression in
0199ff8). We add a Behat test to cover this. Previously this tested that
a non-existent element was not present. The element selector has been
fixed and we now check the user name and picture are present when blind
marking = No, but not present when blind marking = Yes.

Also fixed is CTP-5440 (Plagiarism flag not saved for group) by checking
the allocatable is a user before getting the user picture when re-
rendering the cell following an AJAX update, with the above check for
blind marking. Behat test coverage for this is also added.